moving and formatting
