+++ simplesource-command-kafka/src/main/java/io/simplesource/kafka/internal/streams/EventSourcedStreamsApp.java
@@ -97,9 +97,8 @@ public final class EventSourcedStreamsApp {
         final StreamsBuilder builder = new StreamsBuilder();
         aggregateSetSpec.aggregateConfigMap()
                 .values()
-                .forEach(aggregateSpec -> {
-                    EventSourcedTopology.addTopology(new TopologyContext<>(aggregateSpec), builder);
-                });
+                .forEach(aggregateSpec ->
+                        EventSourcedTopology.addTopology(new TopologyContext<>(aggregateSpec), builder));
         return builder.build();
     }
