+++ simplesource-command-kafka/src/main/java/io/simplesource/kafka/internal/streams/topology/EventSourcedConsumer.java
@@ -6,7 +6,8 @@ import org.apache.kafka.streams.kstream.KStream;
 
 import static io.simplesource.kafka.api.AggregateResources.TopicEntity.command_request;
 
-public final class EventSourcedConsumer {
+final class EventSourcedConsumer {
+
     static <K, C, E, A> KStream<K, CommandRequest<C>> commandRequestStream(TopologyContext<K, C, E, A> ctx, final StreamsBuilder builder) {
         return builder.stream(ctx.topicName(command_request), ctx.commandEventsConsumed());
     }
