+++ simplesource-command-kafka/src/main/java/io/simplesource/kafka/internal/streams/topology/EventSourcedStreams.java
@@ -28,68 +28,7 @@ import static io.simplesource.kafka.api.AggregateResources.StateStoreEntity.aggr
 import static io.simplesource.kafka.api.AggregateResources.StateStoreEntity.command_response;
 import static org.apache.kafka.streams.state.Stores.persistentKeyValueStore;
 
-final class EventSourcedStores {
-
-    static <K, C, E, A> void addStateStores(TopologyContext<K, C, E, A> ctx, final StreamsBuilder builder) {
-
-        final KeyValueStoreBuilder<K, AggregateUpdate<A>> aggregateStoreBuilder = new KeyValueStoreBuilder<>(
-                persistentKeyValueStore(ctx.stateStoreName(aggregate_update)),
-                ctx.serdes().aggregateKey(),
-                ctx.serdes().aggregateUpdate(),
-                Time.SYSTEM);
-        builder.addStateStore(aggregateStoreBuilder);
-    }
-
-    /**
-     * Update the state store with the latest aggregate_update value on successful updates.
-     */
-    static <K, C, E, A> void updateAggregateStateStore(TopologyContext<K, C, E, A> ctx, final KStream<K, AggregateUpdateResult<A>> aggregateUpdateStream) {
-        aggregateUpdateStream.process(() -> new Processor<K, AggregateUpdateResult<A>>() {
-            private KeyValueStore<K, AggregateUpdate<A>> stateStore;
-
-            @Override
-            public void init(final ProcessorContext context) {
-                stateStore = (KeyValueStore<K, AggregateUpdate<A>>) context.getStateStore(ctx.stateStoreName(aggregate_update));
-            }
-
-            @Override
-            public void process(final K readOnlyKey, final AggregateUpdateResult<A> aggregateUpdateResult) {
-                aggregateUpdateResult.updatedAggregateResult().ifSuccessful(
-                        aggregateUpdate -> stateStore.put(
-                                readOnlyKey,
-                                aggregateUpdate));
-            }
-
-            @Override
-            public void close() {
-            }
-        }, ctx.stateStoreName(aggregate_update));
-    }
-
-    static <K, C, E, A> void updateCommandResultStore(TopologyContext<K, C, E, A> ctx, final KStream<K, AggregateUpdateResult<A>> aggregateUpdateStream) {
-        final long retentionMillis = TimeUnit.SECONDS.toMillis(ctx.commandResponseRetentionInSeconds());
-        aggregateUpdateStream
-                .map((k, v) -> KeyValue.pair(
-                        v.commandId(),
-                        v))
-                .groupByKey(ctx.serializedAggregateUpdate())
-                .windowedBy(
-                        TimeWindows
-                                .of(retentionMillis)
-                                .advanceBy(retentionMillis / 3L))
-                .reduce((current, latest) -> latest, materializedWindow(ctx, ctx.stateStoreName(command_response)));
-    }
-
-    private static <K, C, E, A> Materialized<UUID, AggregateUpdateResult<A>, WindowStore<Bytes, byte[]>> materializedWindow(TopologyContext<K, C, E, A> ctx, final String storeName) {
-        return Materialized
-                .<UUID, AggregateUpdateResult<A>, WindowStore<Bytes, byte[]>>as(storeName)
-                .withKeySerde(ctx.serdes().commandResponseKey())
-                .withValueSerde(ctx.serdes().updateResult());
-    }
-}
-
-public final class EventSourcedStreams {
-    private static final Logger logger = LoggerFactory.getLogger(EventSourcedTopology.class);
+final class EventSourcedStreams {
 
     static <K, C, E, A> KStream<K, CommandEvents<E, A>> eventResultStream(TopologyContext<K, C, E, A> ctx, final KStream<K, CommandRequest<C>> commandRequestStream) {
         return commandRequestStream
